# Movegen speedup via king attack info