--- a/src/chess/board.cc
+++ b/src/chess/board.cc
@@ -30,6 +30,7 @@
 #include <cctype>
 #include <cstdlib>
 #include <cstring>
+#include <iostream>
 #include <sstream>
 #include "utils/exception.h"
 
@@ -179,6 +180,24 @@ static const Move::Promotion kPromotions[] = {
     Move::Promotion::Bishop,
     Move::Promotion::Knight,
 };
+
+string DebugStringBB(std::uint64_t bb) {
+  string result;
+  for (int i = 7; i >= 0; --i) {
+    for (int j = 0; j < 8; ++j) {
+      if (bb & (1ULL << (8 * i + j))) {
+        result += '*';
+      } else {
+        result += '.';
+      }
+
+      result += ' ';
+    }
+
+    result += '\n';
+  }
+  return result;
+}
 
 }  // namespace
 
@@ -454,13 +473,13 @@ bool ChessBoard::ApplyMove(Move move) {
 bool ChessBoard::IsUnderAttack(BoardSquare square) const {
   const int row = square.row();
   const int col = square.col();
-  // Check king
+  // Check king.
   {
     const int krow = their_king_.row();
     const int kcol = their_king_.col();
     if (std::abs(krow - row) <= 1 && std::abs(kcol - col) <= 1) return true;
   }
-  // Check Rooks (and queen)
+  // Check rooks (and queens).
   if (kRookAttacks[square.as_int()].intersects(their_pieces_ * rooks_)) {
     for (const auto& direction : kRookDirections) {
       auto dst_row = row;
@@ -478,7 +497,7 @@ bool ChessBoard::IsUnderAttack(BoardSquare square) const {
       }
     }
   }
-  // Check Bishops
+  // Check bishops.
   if (kBishopAttacks[square.as_int()].intersects(their_pieces_ * bishops_)) {
     for (const auto& direction : kBishopDirections) {
       auto dst_row = row;
@@ -496,11 +515,11 @@ bool ChessBoard::IsUnderAttack(BoardSquare square) const {
       }
     }
   }
-  // Check pawns
+  // Check pawns.
   if (kPawnAttacks[square.as_int()].intersects(their_pieces_ * pawns_)) {
     return true;
   }
-  // Check knights
+  // Check knights.
   {
     if (kKnightAttacks[square.as_int()].intersects(their_pieces_ - their_king_ -
                                                    rooks_ - bishops_ -
@@ -511,17 +530,135 @@ bool ChessBoard::IsUnderAttack(BoardSquare square) const {
   return false;
 }
 
-bool ChessBoard::IsLegalMove(Move move, bool was_under_check) const {
+void ChessBoard::GenerateKingAttackInfo(
+    ChessBoard::KingAttackInfo& king_attack_info) const {
+  unsigned num_attackers = 0;
+
+  const int row = our_king_.row();
+  const int col = our_king_.col();
+  // Check king.
+  {
+    const int krow = their_king_.row();
+    const int kcol = their_king_.col();
+    if (std::abs(krow - row) <= 1 && std::abs(kcol - col) <= 1) {
+      king_attack_info.attacking_squares_and_lines_.set(their_king_);
+      num_attackers++;
+    }
+  }
+  // Check rooks (and queens).
+  if (kRookAttacks[our_king_.as_int()].intersects(their_pieces_ * rooks_)) {
+    for (const auto& direction : kRookDirections) {
+      auto dst_row = row;
+      auto dst_col = col;
+      BitBoard attack_line(0);
+      bool possible_pinned_piece_found = false;
+      BoardSquare possible_pinned_piece;
+      while (true) {
+        dst_row += direction.first;
+        dst_col += direction.second;
+        if (!BoardSquare::IsValid(dst_row, dst_col)) break;
+        const BoardSquare destination(dst_row, dst_col);
+        if (our_pieces_.get(destination)) {
+          if (possible_pinned_piece_found) {
+            // No pieces pinned.
+            break;
+          } else {
+            // This is a possible pinned piece;
+            possible_pinned_piece_found = true;
+            possible_pinned_piece = destination;
+          }
+        }
+        if (!possible_pinned_piece_found) {
+          attack_line.set(destination);
+        }
+        if (their_pieces_.get(destination)) {
+          if (rooks_.get(destination)) {
+            if (possible_pinned_piece_found) {
+              // Store the pinned piece.
+              king_attack_info.pinned_pieces_.set(possible_pinned_piece);
+            } else {
+              // Update attacking lines.
+              king_attack_info.attacking_squares_and_lines_ =
+                  king_attack_info.attacking_squares_and_lines_ + attack_line;
+              num_attackers++;
+            }
+          }
+          break;
+        }
+      }
+    }
+  }
+  // Check bishops.
+  if (kBishopAttacks[our_king_.as_int()].intersects(their_pieces_ * bishops_)) {
+    for (const auto& direction : kBishopDirections) {
+      auto dst_row = row;
+      auto dst_col = col;
+      BitBoard attack_line(0);
+      bool possible_pinned_piece_found = false;
+      BoardSquare possible_pinned_piece;
+      while (true) {
+        dst_row += direction.first;
+        dst_col += direction.second;
+        if (!BoardSquare::IsValid(dst_row, dst_col)) break;
+        const BoardSquare destination(dst_row, dst_col);
+        if (our_pieces_.get(destination)) {
+          if (possible_pinned_piece_found) {
+            // No pieces pinned.
+            break;
+          } else {
+            // This is a possible pinned piece;
+            possible_pinned_piece_found = true;
+            possible_pinned_piece = destination;
+          }
+        }
+        if (!possible_pinned_piece_found) {
+          attack_line.set(destination);
+        }
+        if (their_pieces_.get(destination)) {
+          if (bishops_.get(destination)) {
+            if (possible_pinned_piece_found) {
+              // Store the pinned piece.
+              king_attack_info.pinned_pieces_.set(possible_pinned_piece);
+            } else {
+              // Update attacking lines.
+              king_attack_info.attacking_squares_and_lines_ =
+                  king_attack_info.attacking_squares_and_lines_ + attack_line;
+              num_attackers++;
+            }
+          }
+          break;
+        }
+      }
+    }
+  }
+  // Check pawns.
+  BitBoard attacking_pawns =
+      kPawnAttacks[our_king_.as_int()] * their_pieces_ * pawns_;
+  king_attack_info.attacking_squares_and_lines_ =
+      king_attack_info.attacking_squares_and_lines_ + attacking_pawns;
+
+  if (attacking_pawns.as_int()) {
+    num_attackers++;
+  }
+
+  // Check knights.
+  BitBoard attacking_knights =
+      kKnightAttacks[our_king_.as_int()] *
+      (their_pieces_ - their_king_ - rooks_ - bishops_ - (pawns_ * kPawnMask));
+  king_attack_info.attacking_squares_and_lines_ =
+      king_attack_info.attacking_squares_and_lines_ + attacking_knights;
+
+  if (attacking_knights.as_int()) {
+    num_attackers++;
+  }
+
+  king_attack_info.double_check_ = (num_attackers > 1);
+}
+
+bool ChessBoard::IsLegalMove(Move move,
+                             const KingAttackInfo& king_attack_info) const {
   const auto& from = move.from();
   const auto& to = move.to();
-
-  // If we are already under check, also apply move and check if valid.
-  // TODO(mooskagh) Optimize this case
-  if (was_under_check) {
-    ChessBoard board(*this);
-    board.ApplyMove(move);
-    return !board.IsUnderCheck();
-  }
 
   // En passant. Complex but rare. Just apply
   // and check that we are not under check.
@@ -530,6 +667,56 @@ bool ChessBoard::IsLegalMove(Move move, bool was_under_check) const {
     ChessBoard board(*this);
     board.ApplyMove(move);
     return !board.IsUnderCheck();
+  }
+
+  // Check if we are already under check.
+  if (king_attack_info.in_check()) {
+#if 0
+    std::cout << "King is in check." << std::endl;
+    std::cout << "is double check " << king_attack_info.in_double_check()
+              << std::endl;
+    std::cout << "pinned piece: " << king_attack_info.is_pinned_piece(from)
+              << std::endl;
+    std::cout << "Position:" << std::endl << DebugString() << std::endl;
+    std::cout << "Move:" << move.as_string() << std::endl;
+    std::cout << "Attack set:" << std::endl
+              << DebugStringBB(
+                     king_attack_info.attacking_squares_and_lines_.as_int())
+              << std::endl;
+    std::cout << "Pinned pieces:" << std::endl
+              << DebugStringBB(king_attack_info.pinned_pieces_.as_int())
+              << std::endl;
+#endif
+
+    // King movement.
+    if (from == our_king_) {
+      if (king_attack_info.is_attacked_square(to) && !their_pieces_.get(to))
+        return false;  // fast fall-through
+
+      // Slower fallback check.
+      ChessBoard board(*this);
+      board.ApplyMove(move);
+      return !board.IsUnderCheck();
+    }
+
+    if (!king_attack_info.in_double_check()) {
+      // We are not in double check.
+      if (!king_attack_info.is_pinned_piece(from)) {
+        // The piece is free to move. Check if the attacker is captured or
+        // interposed if the piece moves to its destination square.
+        return king_attack_info.is_attacked_square(to);
+      } else {
+        if (!king_attack_info.is_attacked_square(to)) return false;
+
+        // Pinned piece fallback check.
+        ChessBoard board(*this);
+        board.ApplyMove(move);
+        return !board.IsUnderCheck();
+      }
+    } else {
+      // At this stage, no move can resolve the double checks anymore.
+      return false;
+    }
   }
 
   // If it's king's move, check that destination
@@ -579,13 +766,15 @@ bool ChessBoard::IsLegalMove(Move move, bool was_under_check) const {
 }
 
 MoveList ChessBoard::GenerateLegalMoves() const {
-  const bool was_under_check = IsUnderCheck();
+  KingAttackInfo king_attack_info;
+  if (IsUnderCheck()) GenerateKingAttackInfo(king_attack_info);
+
   MoveList move_list = GeneratePseudolegalMoves();
   MoveList result;
   result.reserve(move_list.size());
 
   for (Move m : move_list) {
-    if (IsLegalMove(m, was_under_check)) result.emplace_back(m);
+    if (IsLegalMove(m, king_attack_info)) result.emplace_back(m);
   }
 
   return result;