Various movegen optimizations centered around the introduction of precomputed *king attack info* (attacking lines, pinned pieces, double check detection) that is used to speedup several parts of legal move detection (for all moves).

Unconditional speedup for random backend up to 6% (tested with LTO builds on Linux Fedora 29 64-bit, `gcc 8.2.1`).


The following measurements all share the command line `lc0 benchmark --backend=random --smart-pruning-factor=0 --nodes=1000000 --nncache=1000000`. A 95% confidence interval for the average of the last reported nps value over 25 runs is given for the starting position, the opening position https://lichess.org/editor/rn2kb1r/pp3ppp/2p1pnb1/8/3P1N1q/4B1N1/PPPQ1PPP/R3KB1R_b_KQkq_-_5_9 and the endgame position https://lichess.org/editor/8/2R2P2/1p4K1/6R1/1r6/5rP1/1k6/2q5_w_-_-_0_55 .

The measurements are performed on an Intel Core i5-5300U CPU @ 2.30GHz with 4 cores and L1d cache: 32K, L1i cache: 32K, L2 cache: 256K, L3 cache: 3072K.

For 1 search thread:

position  |  master `6a639b6` |  this PR  | speedup
------ | ----- | ----------- | -----------
start | 208222 ± 448 | 215247 ± 726 | 3.4%
opening | 161321 ± 216 | 168776 ± 335 | 4.6%
endgame | 176833 ± 245 | 186576 ± 283 | 5.5%

For 2 search threads:

position  |  master `6a639b6` |  this PR  | speedup
------ | ----- | ----------- | -----------
start | 271137 ± 830 | 272221 ± 882 | 0.4%
opening | 224128 ± 422 | 227968 ± 551 | 1.7%
endgame | 238111 ± 511 | 243892 ± 758 | 2.4%