@mooskagh 2019-01-05
On my machine chessboard_test with uncommented tests:
master:
```
[ RUN      ] ChessBoard.MoveGenStartingPos
[       OK ] ChessBoard.MoveGenStartingPos (6676 ms)
[ RUN      ] ChessBoard.MoveGenKiwipete
[       OK ] ChessBoard.MoveGenKiwipete (11287 ms)
[ RUN      ] ChessBoard.MoveGenPosition3
[       OK ] ChessBoard.MoveGenPosition3 (868 ms)
[ RUN      ] ChessBoard.MoveGenPosition4
[       OK ] ChessBoard.MoveGenPosition4 (1052 ms)
[ RUN      ] ChessBoard.MoveGenPosition5
[       OK ] ChessBoard.MoveGenPosition5 (5088 ms)
[ RUN      ] ChessBoard.MoveGenPosition6
[       OK ] ChessBoard.MoveGenPosition6 (202 ms)
```

With this PR
```
[ RUN      ] ChessBoard.MoveGenStartingPos
[       OK ] ChessBoard.MoveGenStartingPos (5931 ms)
[ RUN      ] ChessBoard.MoveGenKiwipete
[       OK ] ChessBoard.MoveGenKiwipete (9873 ms)
[ RUN      ] ChessBoard.MoveGenPosition3
[       OK ] ChessBoard.MoveGenPosition3 (739 ms)
[ RUN      ] ChessBoard.MoveGenPosition4
[       OK ] ChessBoard.MoveGenPosition4 (892 ms)
[ RUN      ] ChessBoard.MoveGenPosition5
[       OK ] ChessBoard.MoveGenPosition5 (4413 ms)
[ RUN      ] ChessBoard.MoveGenPosition6
[       OK ] ChessBoard.MoveGenPosition6 (184 ms)
```

Looks good!